Started damage functionality for leap and chomp
Started working on the damage aspects of these abilities.

Need to rework the spline movement functionality so that the boss moves to a start position and then moves along the spline, the aim is have more accurate movement since right now it's kind of hit or miss.

Also need to smooth the transitions between abilities to returning to circle with a lerp and timeline.